1 Page of Front-End
A single front-end sample has been uploaded to demonstrate the basic layout.
This sample includes one tab only, serving as a foundation for the rest of the front-end project.

All necessary CSS and JavaScript files that will be used across the full front-end system have been included in this file to ensure reusability and consistency.